+++ realm/realm-library/src/androidTest/java/io/realm/entities/BacklinksTarget.java
@@ -20,6 +20,10 @@ import io.realm.RealmResults;
 import io.realm.annotations.LinkingObjects;
 
 public class BacklinksTarget extends RealmObject {
+    public static final String CLASS_NAME = "BacklinksTarget";
+    public static final String FIELD_ID = "id";
+    public static final String FIELD_PARENTS = "parents";
+
     private int id;
 
     @LinkingObjects("child")