+++ realm/realm-library/src/main/java/io/realm/OsRealmSchema.java
@@ -34,10 +34,15 @@ import io.realm.internal.Table;
  */
 class OsRealmSchema extends RealmSchema {
     static final class Creator extends RealmSchema {
-        private final Map<String, OsRealmObjectSchema> schema = new HashMap<>();
+        private final Map<String, RealmObjectSchema> schema = new HashMap<>();
 
         @Override
-        public void close() { }
+        public void close() {
+            for (Map.Entry<String, RealmObjectSchema> entry : schema.entrySet()) {
+                entry.getValue().close();
+            }
+            schema.clear();
+        }
 
         @Override
         public RealmObjectSchema get(String className) {
@@ -46,7 +51,7 @@ class OsRealmSchema extends RealmSchema {
         }
 
         @Override
-        public Set<OsRealmObjectSchema> getAll() {
+        public Set<RealmObjectSchema> getAll() {
             return new LinkedHashSet<>(schema.values());
         }
 
@@ -96,14 +101,14 @@ class OsRealmSchema extends RealmSchema {
 
     private final Map<String, RealmObjectSchema> dynamicClassToSchema = new HashMap<>();
 
-    private final long nativePtr;
+    private long nativePtr;
 
     OsRealmSchema(Creator creator) {
-        Set<OsRealmObjectSchema> realmObjectSchemas = creator.getAll();
+        Set<RealmObjectSchema> realmObjectSchemas = creator.getAll();
         long[] schemaNativePointers = new long[realmObjectSchemas.size()];
         int i = 0;
-        for (OsRealmObjectSchema schema : realmObjectSchemas) {
-            schemaNativePointers[i++] = schema.getNativePtr();
+        for (RealmObjectSchema schema : realmObjectSchemas) {
+            schemaNativePointers[i++] = ((OsRealmObjectSchema) schema).getNativePtr();
         }
         this.nativePtr = nativeCreateFromList(schemaNativePointers);
     }
@@ -115,11 +120,10 @@ class OsRealmSchema extends RealmSchema {
     // See BaseRealm uses a StandardRealmSchema, not a OsRealmSchema.
     @Override
     public void close() {
-        Set<OsRealmObjectSchema> schemas = getAll();
-        for (RealmObjectSchema schema : schemas) {
-            schema.close();
+        if (nativePtr != 0L) {
+            nativeClose(nativePtr);
+            nativePtr = 0L;
         }
-        nativeClose(nativePtr);
     }
 
     /**
@@ -140,13 +144,8 @@ class OsRealmSchema extends RealmSchema {
      * @return the set of all classes in this Realm or no RealmObject classes can be saved in the Realm.
      */
     @Override
-    public Set<OsRealmObjectSchema> getAll() {
-        long[] ptrs = nativeGetAll(nativePtr);
-        Set<OsRealmObjectSchema> schemas = new LinkedHashSet<>(ptrs.length);
-        for (int i = 0; i < ptrs.length; i++) {
-            schemas.add(new OsRealmObjectSchema(ptrs[i]));
-        }
-        return schemas;
+    public Set<RealmObjectSchema> getAll() {
+        throw new UnsupportedOperationException();
     }
 
     /**
@@ -214,6 +213,4 @@ class OsRealmSchema extends RealmSchema {
     static native long nativeCreateFromList(long[] objectSchemaPtrs);
 
     static native void nativeClose(long nativePtr);
-
-    static native long[] nativeGetAll(long nativePtr);
 }